first

+++ src/main/java/com/example/demo/model/Product.java
@@ -1,19 +1,19 @@
 package com.example.demo.model;
 
 public class Product {
-    private Long id;
+    private int id;
     private String name;
 
-    public Product(Long id, String name) {
+    public Product(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
